move blazor server host-related files into dedicated feature directory
