+++ WoWsShipBuilder.Web/Common/App.razor
@@ -17,5 +17,5 @@
 
 @code
 {
-    private List<Assembly> assemblies = new() { typeof(ShipStats).Assembly };
+    private readonly List<Assembly> assemblies = new() { typeof(ShipStats).Assembly };
 }
