+++ WoWsShipBuilder.Web/Features/Host/Error.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model WoWsShipBuilder.Web.Common.ErrorModel
+@model WoWsShipBuilder.Web.Features.Host.ErrorModel
 
 <!DOCTYPE html>
 <html lang="en">
