+++ WoWsShipBuilder.Web/Features/Host/Error.cshtml.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace WoWsShipBuilder.Web.Common;
+namespace WoWsShipBuilder.Web.Features.Host;
 
 [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
 [IgnoreAntiforgeryToken]
@@ -11,13 +11,6 @@ public class ErrorModel : PageModel
     public string? RequestId { get; set; }
 
     public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-
-    private readonly ILogger<ErrorModel> logger;
-
-    public ErrorModel(ILogger<ErrorModel> logger)
-    {
-        this.logger = logger;
-    }
 
     public void OnGet()
     {
