+++ WoWsShipBuilder.Web/Features/Host/_Host.cshtml
@@ -1,9 +1,9 @@
 ﻿@page "/"
 @using WoWsShipBuilder.Web.Common
-@namespace WoWsShipBuilder.Web.Pages
+@namespace WoWsShipBuilder.Web.Features.Host
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @{
-    Layout = "_Layout";
+    Layout = "_Layout.cshtml";
 }
 
 <component type="typeof(App)" render-mode="Server"/>
