+++ WoWsShipBuilder.Web/Features/Host/_Layout.cshtml
@@ -1,7 +1,7 @@
 ﻿@using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Mvc.TagHelpers
 @using WoWsShipBuilder.Infrastructure.Utility
-@namespace WoWsShipBuilder.Web.Pages
+@namespace WoWsShipBuilder.Web.Features.Host
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 
 <!DOCTYPE html>
